Add a stylesheet with a 2 column layout
This is responsive, so that the sidebar turns into a footer if the
screen is too narrow. If rendering as a sidebar we want to make sure
there is enough room to display long page titles, and enough room to
render articles containing 80 character wide code blocks.

I've also limited the width of the article to make text blocks easier to
read, so on very wide screens the extra space will fill the sidebar.