Tweak CSS
Firstly, reduce font size - it was really big and it looked weird after
changing the font.

Secondly, prevent the sidebar from growing to consume space. At the
moment, this means there is unused space on the right hand side when
the screen is wide enough. In a future commit, I'll change this so
that the article column grows, but there'll be a separate text container
within that to restrict line length.