Stop using wrap-reverse
Previously, I used flex-wrap: wrap-reverse so that the nav turned into a
footer on narrow screens.

However, this creates very strange behaviour when navigating via
keyboard or assistive tech, because the DOM order and focus order do not
match the visual order.

I think it's better to just not use this. Instead, we can put the nav on
the right hand side so the order is the same regardless of whether the
content is laid out left to right or top to bottom.